+++ lib/ui/explore/ExploreBuildingDetailPanel.dart
@@ -16,8 +16,9 @@ import 'package:rokwire_plugin/utils/utils.dart';
 class ExploreBuildingDetailPanel extends StatefulWidget {
   final Building? building;
   final String? buildingNumber;
+  final ExploreSelectLocationBuilder? selectLocationBuilder;
 
-  ExploreBuildingDetailPanel({Key? key, this.building, this.buildingNumber}) : super(key: key);
+  ExploreBuildingDetailPanel({Key? key, this.building, this.buildingNumber, this.selectLocationBuilder }) : super(key: key);
 
   @override
   State<StatefulWidget> createState() => _ExploreBuildingDetailPanelState();
@@ -93,6 +94,7 @@ class _ExploreBuildingDetailPanelState extends State<ExploreBuildingDetailPanel>
       _buildLocation(),
       _buildShare(),
       // _buildFloorPlansAndAmenities(),
+      _buildSelectLocation(),
     ]);
 
   Widget _buildTitle() =>
@@ -156,6 +158,12 @@ class _ExploreBuildingDetailPanelState extends State<ExploreBuildingDetailPanel>
         ),
       ),
     );
+
+  Widget _buildSelectLocation() {
+    Widget? selectorWidget = widget.selectLocationBuilder?.call(context, ExploreSelectLocationContext.detail, explore: _building);
+    return (selectorWidget != null) ? Padding(padding: EdgeInsets.only(top: 32), child: selectorWidget) : Container();
+  }
+
 
   Widget _buildLoadingContent() => Center(child:
     Padding(padding: EdgeInsets.zero, child: