+++ lib/ui/explore/ExploreListPanel.dart
@@ -44,8 +44,9 @@ class ExploreListPanel extends StatefulWidget with AnalyticsInfo {
   final List<Explore>? explores;
   final ExploreMapType? exploreMapType;
   final Position? initialLocationData;
+  final ExploreSelectLocationBuilder? selectLocationBuilder;
 
-  ExploreListPanel({this.explores, this.exploreMapType, this.initialLocationData});
+  ExploreListPanel({ this.explores, this.exploreMapType, this.initialLocationData, this.selectLocationBuilder });
 
   @override
   _ExploreListPanelState createState() =>
@@ -142,14 +143,18 @@ class _ExploreListPanelState extends State<ExploreListPanel> implements Notifica
     }
     else {
       return Padding(padding: EdgeInsets.only(top: 16, bottom: isLast ? 16 : 0), child:
-        ExploreCard(explore: explore, locationData: widget.initialLocationData, onTap: () => _onTapExplore(explore)),
+        ExploreCard(explore: explore,
+          locationData: widget.initialLocationData,
+          selectLocationBuilder: widget.selectLocationBuilder,
+          onTap: () => _onTapExplore(explore)
+        ),
       );
     }
   }
 
   void _onTapExplore(Explore explore) {
     Analytics().logSelect(target: explore.exploreTitle);
-    explore.exploreLaunchDetail(context, initialLocationData: widget.initialLocationData,);
+    explore.exploreLaunchDetail(context, initialLocationData: widget.initialLocationData, selectLocationBuilder: widget.selectLocationBuilder);
   }
 
   void _onTapMTDStop(MTDStop? stop) {