+++ CHANGELOG.md
@@ -5,8 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [6.1.12] - 2024-10-15
 ### Changed
 - Cleaned up shared explore UI [#4363](https://github.com/rokwire/illinois-app/issues/4363).
+- Updated SafeWalks UI [#4363](https://github.com/rokwire/illinois-app/issues/4363).
+### Fixed
+- Fixed map location selection for group markers [#4363](https://github.com/rokwire/illinois-app/issues/4363).
+### Added
+- Added building search capability in location selection [#4363](https://github.com/rokwire/illinois-app/issues/4363).
 
 ## [6.1.11] - 2024-10-09
 ### Added
@@ -72,6 +78,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [6.1.0] - 2024-07-22
 ### Added
 - Created Public Surveys UI / round 0 [#4243](https://github.com/rokwire/illinois-app/issues/4243).
+
+## [6.0.51] - 2024-10-15
+### Fixed
+- Fixed Poll UI issues [#4414](https://github.com/rokwire/illinois-app/issues/4414).
 
 ## [6.0.50] - 2024-10-03
 ### Fixed